+++ packages/design-system/src/Avatar/Avatar.stories.tsx
@@ -18,4 +18,5 @@ AvatarStory.storyName = "Avatar";
 AvatarStory.args = {
   image: "https://picsum.photos/200/300",
   label: "Avatar",
+  isTooltipEnabled: true,
 };