+++ packages/design-system/src/Callout/Callout.styles.tsx
@@ -40,15 +40,20 @@ export const StyledCallout = styled.div<{
   border-radius: var(--ads-v2-border-radius);
   padding: var(--ads-v2-spaces-3) var(--ads-v2-spaces-4);
 
+  // to ensure that small window sizes don't reduce the background size of callout
+  flex-shrink: 0;
+
   background-color: var(--callout-color-background);
 
   ${({ isClosed }) => isClosed && `display: none;`}
 `;
 
 export const StyledCloseButton = styled(Button)`
-  margin-left: auto;
-  cursor: pointer;
-  min-width: fit-content;
+  && {
+    margin-left: auto;
+    cursor: pointer;
+    min-width: fit-content;
+  }
 `;
 
 export const StyledLinks = styled.div`