+++ packages/design-system/src/SegmentedControl/SegmentedControl.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import clsx from "classnames";
 
 import { SegmentedControlProps } from "./SegmentedControl.types";
@@ -12,6 +12,10 @@ import { Text } from "Text";
 import {
   SegmentedControlClassName,
   SegmentedControlSegmentClassName,
+  SegmentedControlSegmentEndIconClassName,
+  SegmentedControlSegmentStartIconClassName,
+  SegmentedControlSegmentTextClassName,
+  SegmentedControlSegmentValueClassName,
   SegmentedControlSegmentsContainerClassName,
 } from "./SegmentedControl.constants";
 
@@ -23,23 +27,62 @@ const SegmentedControl = React.forwardRef(
       isFullWidth = true,
       onChange,
       options,
+      value: controlledValue,
       ...rest
     } = props;
-    const [selectedValue, setSelectedValue] = useState(defaultValue);
+    const segmentRefs: Array<HTMLSpanElement | null> = [];
+    const [selectedValue, setSelectedValue] = useState(
+      controlledValue ?? defaultValue,
+    );
+    const [focusedIndex, setFocusedIndex] = useState<number>(0);
 
-    const handleOnChange = (value: string) => {
-      setSelectedValue(value);
-      onChange && onChange(value);
+    useEffect(() => {
+      if (controlledValue !== undefined) {
+        setSelectedValue(controlledValue);
+      }
+    }, [controlledValue]);
+
+    const handleKeyDown = (e: React.KeyboardEvent, index: number) => {
+      if (!segmentRefs.length) return;
+
+      switch (e.key) {
+        case "ArrowRight":
+        case "Right":
+          const rightIndex = index === options.length - 1 ? 0 : index + 1;
+          segmentRefs[rightIndex]?.focus();
+          setFocusedIndex(rightIndex);
+          break;
+
+        case "ArrowLeft":
+        case "Left":
+          const leftIndex = index === 0 ? options.length - 1 : index - 1;
+          segmentRefs[leftIndex]?.focus();
+          setFocusedIndex(leftIndex);
+          break;
+
+        case "Enter":
+        case " ":
+          handleOnChange(options[index].value, true);
+          break;
+      }
+    };
+
+    const handleOnChange = (value: string, isUpdatedViaKeyboard = false) => {
+      if (controlledValue === undefined) {
+        setSelectedValue(value);
+      }
+      onChange && onChange(value, isUpdatedViaKeyboard);
     };
 
     return (
       <StyledSegmentedControl
         className={clsx(SegmentedControlClassName, className)}
         isFullWidth={isFullWidth}
+        onBlur={() => setFocusedIndex(0)}
         ref={ref}
         {...rest}
       >
-        {options.map((option) => {
+        {options.map((option, index) => {
           const {
             endIcon,
             isDisabled: isSegmentDisabled,
@@ -55,30 +98,44 @@ const SegmentedControl = React.forwardRef(
               data-selected={selectedValue === value}
               key={value}
               onClick={() => !isSegmentDisabled && handleOnChange(value)}
+              onKeyDown={(event) => handleKeyDown(event, index)}
+              ref={(input) => segmentRefs.push(input)}
+              tabIndex={index === focusedIndex ? 0 : -1}
             >
               <StyledSegment
-                className={SegmentedControlSegmentClassName}
+                className={clsx(SegmentedControlSegmentClassName)}
                 data-selected={selectedValue === value}
+                data-value={value}
               >
                 {/* if icon name is passed */}
                 {startIcon && typeof startIcon === "string" && (
-                  <Icon name={startIcon} size="md" />
-                )}
-                {/* if icon component is passed */}
-                {startIcon && typeof startIcon !== "string" && (
-                  <Icon size="md">{startIcon}</Icon>
+                  <Icon
+                    className={SegmentedControlSegmentStartIconClassName}
+                    name={startIcon}
+                    size="md"
+                  />
                 )}
 
                 {/* Label */}
-                {label && <Text kind="body-m">{label}</Text>}
+                {label && (
+                  <Text
+                    className={clsx(
+                      SegmentedControlSegmentTextClassName,
+                      SegmentedControlSegmentValueClassName + value,
+                    )}
+                    kind="body-m"
+                  >
+                    {label}
+                  </Text>
+                )}
 
                 {/* if icon name is passed */}
                 {endIcon && typeof endIcon === "string" && (
-                  <Icon name={endIcon} size="md" />
-                )}
-                {/* if icon component is passed */}
-                {endIcon && typeof endIcon !== "string" && (
-                  <Icon size="md">{endIcon}</Icon>
+                  <Icon
+                    className={SegmentedControlSegmentEndIconClassName}
+                    name={endIcon}
+                    size="md"
+                  />
                 )}
               </StyledSegment>
             </StyledControlContainer>