+++ packages/design-system/src/Tab/Tab.styles.tsx
@@ -7,16 +7,18 @@ const Variables = css`
 `;
 
 export const StyledTabs = styled(RadixTabs.Root)`
+  max-height: fit-content;
   width: 100%;
 `;
 
 export const StyledTabsList = styled(RadixTabs.List)`
   display: flex;
+  background-color: var(--ads-v2-color-bg);
   border-bottom: 1px solid var(--ads-v2-color-border);
   display: flex;
   gap: var(--ads-v2-spaces-4);
   width: 100%;
-  overflow: scroll;
+  overflow: auto;
   padding: var(--ads-v2-spaces-1);
 `;